requirements.txt
aiogram==3.4.1
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
apscheduler==3.10.4
pytz==2023.4
